Added flooring, improved mapping color
Firstly, the ability to add floors was massively improved, and now the map has a nice dark cyan color. 

Secondly, there was an issue with serde_json where it would interpret the array given to it into a tuple of floating points, which is not what I wanted at all lol. So now instead, I moved all the floating point work into the custom.js file, where it will just turn the RGBA colors into RGBA but in a float between 0 and 1, where it is interpreted correctly.

+++ tiled/map_objects.tsx
@@ -14,8 +14,8 @@
  <tile id="1">
   <properties>
    <property name="alpha" type="float" value="255"/>
-   <property name="blue" type="float" value="0"/>
-   <property name="green" type="float" value="0"/>
+   <property name="blue" type="int" value="0"/>
+   <property name="green" type="int" value="0"/>
    <property name="player_collidable" type="bool" value="true"/>
    <property name="player_spawn" type="bool" value="false"/>
    <property name="red" type="float" value="255"/>
@@ -24,7 +24,7 @@
  <tile id="2">
   <properties>
    <property name="alpha" type="float" value="255"/>
-   <property name="blue" type="float" value="0"/>
+   <property name="blue" type="int" value="0"/>
    <property name="green" type="float" value="255"/>
    <property name="player_collidable" type="bool" value="true"/>
    <property name="player_spawn" type="bool" value="false"/>
